Extract copy-and-pasted function
