+++ ceno_utils/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+edition.workspace = true
+license.workspace = true
+name = "ceno_utils"
+version.workspace = true
+
+[dependencies]
