+++ ceno_utils/src/lib.rs
@@ -1,0 +1,12 @@
+// split single u64 value into W slices, each slice got C bits.
+// all the rest slices will be filled with 0 if W x C > 64
+pub fn u64vec<const W: usize, const C: usize>(x: u64) -> [u64; W] {
+    assert!(C <= 64);
+    let mut x = x;
+    let mut ret = [0; W];
+    for item in &mut ret {
+        *item = x & ((1 << C) - 1);
+        x >>= C;
+    }
+    ret
+}
