+++ singer/src/lib.rs
@@ -20,7 +20,7 @@ pub mod instructions;
 pub mod scheme;
 #[cfg(test)]
 pub mod test;
-pub use utils::u64vec;
+pub use ceno_utils::u64vec;
 mod utils;
 
 // Process sketch:
